+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Release 0.14.0 (unreleased)
 * Fix ``dfetch add`` crashing with a ``ValueError`` when the remote URL has a trailing slash (#1137)
 * Fix unhelpful error message when a metadata file is malformed (#1145)
 * Fix arbitrary file write via malicious tar/zip symlink (#0)
+* Prevent ssh command injection (#0)
 
 Release 0.13.0 (released 2026-03-30)
 ====================================